+++ src/client/java/minicraft/level/tile/farming/WheatTile.java
@@ -24,7 +24,7 @@ public class WheatTile extends CropTile {
 	public void render(Screen screen, Level level, int x, int y) {
 		int age = (level.getData(x, y) >> 3) & maxAge;
 		Tiles.get("Farmland").render(screen, level, x, y);
-		int stage = (int) (age / maxAge * 5f);
+		int stage = (int) ((float) age / maxAge * 5);
 		screen.render(x * 16, y * 16, spritStages[stage]);
 	}
 }
